Newest version. Still testing\improving. (Working)
This has added the ability for multi-dimensional arrays. Working on the best way to query them, currently works, but not the most effective. Many of the functions need to be cleaned up, to make them more efficient. Still adding more functions. This is getting close to a product I am happy fully pushing out!